requirements.txt
paho-mqtt>=1.6.0\nclick>=8.0.0\n